Fixes #502 Added --verbose flag (#527)
Adds --verbose flag

Fixes #502

+++ cli/popper/gha.py
@@ -13,6 +13,7 @@ import sys
 class Workflow(object):
     """A GHA workflow.
     """
+
     def __init__(self, wfile, workspace):
         if not wfile:
             if os.path.isfile("main.workflow"):
@@ -163,21 +164,21 @@ class Workflow(object):
 
             cloned.add('{}/{}'.format(user, repo))
 
-    def instantiate_runners(self):
+    def instantiate_runners(self, verbose=False):
         """Factory of ActionRunner instances, one for each action"""
         for _, a in self.wf['action'].items():
             if 'docker://' in a['uses']:
                 a['runner'] = DockerRunner(a, self.workspace,
-                                           self.env, self.timeout)
+                                           self.env, self.timeout, verbose)
                 continue
 
             if './' in a['uses']:
                 if os.path.exists(os.path.join(a['uses'], 'Dockerfile')):
                     a['runner'] = DockerRunner(a, self.workspace,
-                                               self.env, self.timeout)
+                                               self.env, self.timeout, verbose)
                 else:
                     a['runner'] = HostRunner(a, self.workspace,
-                                             self.env, self.timeout)
+                                             self.env, self.timeout, verbose)
                 continue
 
             dockerfile_path = os.path.join(a['repo_dir'], a['action_dir'],
@@ -185,17 +186,17 @@ class Workflow(object):
 
             if os.path.exists(dockerfile_path):
                 a['runner'] = DockerRunner(a, self.workspace,
-                                           self.env, self.timeout)
+                                           self.env, self.timeout, verbose)
             else:
                 a['runner'] = HostRunner(a, self.workspace,
-                                         self.env, self.timeout)
+                                         self.env, self.timeout, verbose)
 
-    def run(self, action_name=None, reuse=False):
+    def run(self, action_name=None, reuse=False, verbose=False):
         """Run the pipeline or a specific action"""
         os.environ['WORKSPACE'] = self.workspace
 
         self.download_actions()
-        self.instantiate_runners()
+        self.instantiate_runners(verbose)
 
         if action_name:
             self.wf['action'][action_name]['runner'].run(reuse)
@@ -224,11 +225,13 @@ class Workflow(object):
 class ActionRunner(object):
     """An action runner.
     """
-    def __init__(self, action, workspace, env, timeout):
+
+    def __init__(self, action, workspace, env, timeout, verbose):
         self.action = action
         self.workspace = workspace
         self.env = env
         self.timeout = timeout
+        self.verbose = verbose
 
         if not os.path.exists(self.workspace):
             os.makedirs(self.workspace)
@@ -251,32 +254,39 @@ class ActionRunner(object):
         out_fname = os.path.join(self.log_path, log_tag + '.out')
         err_fname = os.path.join(self.log_path, log_tag + '.err')
 
-        with open(out_fname, "w") as outf, open(err_fname, "w") as errf:
-            p = subprocess.Popen(cmd, stdout=outf, stderr=errf, shell=True,
-                                 preexec_fn=os.setsid)
+        if self.verbose:
+            with open(out_fname, "a") as outf, open(err_fname, "a") as errf:
+                rc = pu.exec_cmd_util(cmd, True, True, outf, errf)[0]
+            pu.info('\n')
+            return rc
+        else:
+            with open(out_fname, "w") as outf, open(err_fname, "w") as errf:
+                p = subprocess.Popen(cmd, stdout=outf, stderr=errf, shell=True,
+                                     preexec_fn=os.setsid)
 
-            while p.poll() is None:
+                while p.poll() is None:
 
-                if self.timeout != 0.0 and time.time() > time_limit:
-                    os.killpg(os.getpgid(p.pid), signal.SIGTERM)
-                    sys.stdout.write(' time out!')
-                    break
+                    if self.timeout != 0.0 and time.time() > time_limit:
+                        os.killpg(os.getpgid(p.pid), signal.SIGTERM)
+                        sys.stdout.write(' time out!')
+                        break
 
-                if sleep_time < 30:
-                    sleep_time *= 2
+                    if sleep_time < 30:
+                        sleep_time *= 2
 
-                for i in range(int(sleep_time)):
-                    if i % 10 == 0:
-                        pu.info('.')
-                time.sleep(sleep_time)
+                    for i in range(int(sleep_time)):
+                        if i % 10 == 0:
+                            pu.info('.')
+                    time.sleep(sleep_time)
 
-        pu.info('\n')
-        return p.poll()
+            pu.info('\n')
+            return p.poll()
 
 
 class DockerRunner(ActionRunner):
-    def __init__(self, action, workspace, env, timeout):
-        super(DockerRunner, self).__init__(action, workspace, env, timeout)
+    def __init__(self, action, workspace, env, timeout, verbose):
+        super(DockerRunner, self).__init__(action, workspace, env,
+                                           timeout, verbose)
         self.cid = self.action['name'].replace(' ', '_')
 
     def run(self, reuse):
@@ -313,7 +323,7 @@ class DockerRunner(ActionRunner):
             pu.fail('Action {} failed!\n'.format(self.action['name']))
 
     def docker_exists(self):
-        cmd_out = pu.exec_cmd('docker ps -a')
+        cmd_out = pu.exec_cmd('docker ps -a', self.verbose)
 
         if self.cid in cmd_out:
             return True
@@ -321,7 +331,8 @@ class DockerRunner(ActionRunner):
         return False
 
     def docker_rm(self):
-        pu.exec_cmd('docker rm {}'.format(self.cid), ignoreerror=True)
+        pu.exec_cmd('docker rm {}'.format(self.cid), self.verbose,
+                                          ignoreerror=True)
 
     def docker_create(self, img):
         env_vars = self.action.get('env', {})
@@ -352,7 +363,7 @@ class DockerRunner(ActionRunner):
             self.action['name'], img, ' '.join(self.action.get('args', '')))
         )
 
-        pu.exec_cmd(docker_cmd)
+        pu.exec_cmd(docker_cmd, self.verbose)
 
     def docker_start(self):
         pu.info('[{}] docker start '.format(self.action['name']))
@@ -362,17 +373,18 @@ class DockerRunner(ActionRunner):
 
     def docker_pull(self, img):
         pu.info('[{}] docker pull {}\n'.format(self.action['name'], img))
-        pu.exec_cmd('docker pull {}'.format(img))
+        pu.exec_cmd('docker pull {}'.format(img), self.verbose)
 
     def docker_build(self, tag, path):
         cmd = 'docker build -t {} {}'.format(tag, path)
         pu.info('[{}] {}\n'.format(self.action['name'], cmd))
-        pu.exec_cmd(cmd)
+        pu.exec_cmd(cmd, self.verbose)
 
 
 class HostRunner(ActionRunner):
-    def __init__(self, action, workspace, env, timeout):
-        super(HostRunner, self).__init__(action, workspace, env, timeout)
+    def __init__(self, action, workspace, env, timeout, verbose):
+        super(HostRunner, self).__init__(action, workspace, env,
+                                         timeout, verbose)
 
     def run(self, reuse=False):
         cmd = self.action.get('runs', ['entrypoint.sh'])
